team: fix handling of D-Bus name watch events
When a device is activated any existing teamd instance is killed. But
since commit 28274495d61c ("device/team: always try to connect to
teamd in update_connection()") the disappearing of the D-Bus name
owner always triggers an automatic restart of the instance in
teamd_dbus_vanished() if the name was previously owned. This new
instance conflicts with the instance we're about to start.

Instead, restore the previous behavior of restarting teamd only if
there is an activation in progress and use @tdc as a flag. This also
means that update_connection() should not modify the value of @tdc.

Fixes: 28274495d61cb434272e0e2307dfd824ee78de8e