device: rework the state transition after IP configuration
Unify the two check_ip_done() and check_ip_failed() functions into a
single one to have all the state transition logic in a single place.

The main difference in behavior introduced by this patch is that now
we wait the termination of the other IP configuration stage when there
is a disabled/ignore method, instead of going immediately to
IP_CHECK. That was a regression introduced by commit 553717bb1c9.

Fixes: 553717bb1c9ed31be0fab85bc37f6823dc8ab480