device: change assuming connections behavior and keep external devices unmanaged (bgo#746440)
Until now, NetworkManager tries to "assume" connections on externally
configured devices.

This behavior is fundamentally wrong because "assuming" has two
contradicting meanings:

a) for one, it means not to touch externally configured devices that
  are not supposed to be managed by NetworkManager. NetworkManager goes
  through   great lengths trying not to mess with such devices, most
  notably software devices like a bridge. NetworkManager does this, to
  present the interface as up and configured, although this is a
  pretend-only mode and the device is not managed by NetworkManager.
b) it also has the meaning during restart, where NetworkManager leaves
  interfaces up during exit and tries to take them over on start. In
  this case, NetworkManager wants to manage the device, but it wants to
  take it over non-destructively, without breaking connectivity.

These are fundamentally different things, because for a) NetworkManager should
not touch the interface at all, while for b) it very much should touch it,
for example to extend DHCP leases and renew addresses.

Until now, this was done by generating an in-memory connection and trying to
match it to any of it's existing setting profiles. If a match is found, that
connection will be assumed, otherwise the generated connection is added as
in-memory connection.
Other then that, there is little difference between whether
NetworkManager is able to match an existing connection or whether
it uses the generated connections. In some cases it affects whether
to lean more towards behavior a) or b). But of courese, the match itself
is only a guess and NetworkManager doesn't make up it's mind whether it wants
to manage the device or not.

Fix that. Now, NetworkManager decides early what to do and whether to manage
the device (b) or not (a). For one, NetworkManager will store the state of the
devices during exit in /var/run/NetworkManager/devices. Thus, after a restart,
NetworkManager knows whether it managed the device before or not. In case
NetworkManager is started the first time since boot, the decision is made
based on user-configuration, UDEV, and default behaviors. Default behaviors
are for example that existing software devices are unmanaged by default,
while real devices default to be managed.

Therefore, the case a), really becomes having external devices as unmanaged.
NetworkManager will no longer pretend a connection is active. This for
example means that no dispatcher events will run for previously assumed
devices. Such a pretend-only, active state could be reintroduced later.

If NetworkManager decides to manage the device (b), it will now move
the device through states unavailable/disconnected, and eventually
autoconnect will start as usual. The graceful take-over is done
only when the run-state file indicates a connection that was active
previously on the device. In that case, NetworkManager will assume
that connection. But once full activation state is reached, an assumed
connection is not any different from a regular
activation/autoactivation.

This of course changes existing behavior in many ways. This is only
a first commit, surely there are many follow up issues.

A lot of code becames now unused. This will be removed in later commits
to keep the diff smaller.

https://bugzilla.gnome.org/show_bug.cgi?id=746440
https://bugzilla.redhat.com/show_bug.cgi?id=1394579