libnm-core: define known route attribute names and validation functions
This adds definition of a set of known route option attributes to
libnm-core and a couple of functions to validate their content.

nm_ip_route_attribute_get_valid_type() returns the valid type for a
known attribute and is useful to parse the string representation of an
attribute in a generic way.

nm_ip_route_attribute_validate() performs the validation of the
attribute type and, in case of a formatted string attribute, of its
content.